fix(offers): transferred models to files
