+++ offers/admin.py
@@ -1,8 +1,11 @@
 from django.contrib import admin
 from django.contrib.admin import register, ModelAdmin
 
-from .models import Advertisement, Category, SubCategory, Image, Attribute
-
+from offers.models.advertisement import Advertisement
+from offers.models.attribute import Attribute
+from offers.models.category import Category
+from offers.models.image import Image
+from offers.models.subcategory import SubCategory
 
 # todo переписать через классы
 admin.site.register(Advertisement)
