+++ offers/models/category.py
@@ -1,0 +1,14 @@
+from django.db import models
+
+
+class Category(models.Model):
+    title = models.CharField(max_length=50, verbose_name="Название")
+    date_create = models.DateTimeField(auto_now_add=True, verbose_name="Дата создания")
+    date_update = models.DateTimeField(auto_now=True, verbose_name="Дата редактирования")
+
+    def __str__(self):
+        return self.title
+
+    class Meta:
+        verbose_name = "Категория"
+        verbose_name_plural = "Категория"
